pyproject.toml
[tool.poetry]
name = "event-propagator"
version = "1.0.0"
description = "Event Propagator service - periodically sends predefined event payloads to a consumer"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.9.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
mypy = "^1.7.0"
ruff = "^0.1.0"
black = "^23.12.0"
pdoc = "^14.1.0"
deptry = ">=0.12.0, <0.14.0"

[tool.poetry.scripts]
propagator = "src.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--strict-markers -ra"
markers = [
    "asyncio: marks tests as async (deselect with '-m \"not asyncio\"')",
    "integration: marks tests as integration tests",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "aiohttp.*",
]
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"
lint.select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "PIE", "SIM"]
lint.ignore = ["E501"]  # Line too long (handled by black)

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-single-line = false

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/site-packages/*"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "^\\s*pass\\s*$",
    "^\\s*\\.\\.\\.\\s*$",
]
fail_under = 85

precision = 2

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.deptry.per_rule_ignores]
DEP002 = ["asyncpg", "pydantic-settings"]


[tool.deptry.package_module_name_map]
"pydantic-settings" = "pydantic_settings"
"pytest-cov" = "pytest_cov"